build: Cache information previously built in settings
All information calculated in settings should be pre-calculated and cached.

This massively reduces time to start the app which is
a) great for devs
b) more importantly, means Heroku won't time us out on start up and crash the whole site.

This commit just has the code to work with this, for ease of reviewing