build: Cache information - store actual information
Storing this in git is the easy option.

If we didn't we would have to run the cache command on Heroku build stage, which seems to need customising a build pack to achieve.

Besides, this way, when reviewing P.R.'s in future that change spreadsheets, there will be some useful data - at the moment reviews are hard because git can't diff binary files well.