data-quality-report: Start charts of listing projects with most issues, per priority
Remove type variables from 2 other DQR views - they were not used and were being set to an unexpected value

Only link user to level 0 priority issues for now - easy to add other links later.
This is becasue currently the data in the other levels is not very good or does not exist.