worker: Move some of update data tasks to only run via worker
As update data tasks are run in a web request, we must make sure they do not run for too long.

So ones that do not to be updated instantly (tagged "low priority") to run via a worker

Currently only project data quality report is done this way. There are other possibilities we could look at later.

But this commit basically means that https://github.com/INDIGO-Initiative/database-app/pull/67 does not add any extra work to a web request.

Also:
New tasks with clearer names for organisations and funds