refactor: Seperate functions for creating / reading data from spreadsheet forms
Previously, we had mixed up data that was intended to be used for other purposes and forms together.

This caused problems as the best form of data for spreadsheet forms may not be the same as for other purposes.

Now, we have a clear set of functions all in indigo/spreadsheetforms.py that
* convert data into data suitable for putting into a form
* extracting edits from form data

Operations here generally have to be symetric;
eg - if you add something in one of those 2 steps you have to remove it in the other one
With this in mind, having them in the same file makes it easy to see the steps together

The ability to do different data processing for forms is not actually used in this commit, but it will be used in the next one and future ones.
(See the history of indigo/spreadsheetforms.py to see if I was right!)

This may also have fixed an error; when downloading some project spreadsheets, indigo.processdata.add_other_records_to_project was being called twice
1. When caching data on project model
2. When preparing data to be put in spreadsheet