standard: Add new data-standard-impact-wayfinder
The assesment resource schema files have moved to a new repository.

Add it as a sub module (like the current data-standard) and change code to look there.

After this commit, assessment_resource.json will be processed twice,
first from the old repository then from the new one.
This is a bit wasteful but is harmless.

When the data-standard submodule is updated to a newer version without
assessment_resource.json this will stop happening.